SKFP-146: make deletePhantomSets int instead of long

+++ src/main/java/bio/overture/riff/controller/RiffController.java
@@ -98,8 +98,8 @@ public class RiffController {
   }
 
   @DeleteMapping("/delete-phantom-sets")
-  public long deletePhantomSets(@RequestHeader(value = HttpHeaders.AUTHORIZATION) final String accessToken) {
-    long result;
+  public int deletePhantomSets(@RequestHeader(value = HttpHeaders.AUTHORIZATION) final String accessToken) {
+    int result;
     val user = jwtFacade.getUser();
     if (user.isPresent()) {
       result = service.deletePhantomSets(user.get());
